src/requirements.txt
Pillow==11.1.0
legacy-cgi==2.6.2
requests==2.32.3
tqdm>=4.67.1
shapely>=2.0.7